b/requirements.txt
python-dotenv
langchain
langchain_openai
pypdf
faiss-cpu
Flask
langchain-core
langchain-community
streamlit